完成切換書本頁面時能自動存檔，打好的內容會自動存到s3 對應檔案
將 current_target.js 合併迴 text_editor 因為切換頁面自動存檔功能需要配合這段code
但是codemirror 的內容只能在 text_editor.js 拿到
