+++ app/javascript/stylesheets/text_editor/index.js
@@ -1,5 +1,3 @@
-import "./current_target.js"
-
 
 import"./add_chapter_and_section.js"
